+++ src/eval/values.c
@@ -4,11 +4,11 @@
 
 int PHASE[PIECE_COUNT] = {0, 1, 1, 2, 4, 0};
 
-const int VALUE_MG[6] = {82, 337, 365, 477, 1025, 0};
-const int VALUE_EG[6] = {94, 281, 297, 512, 936, 0};
+const int16_t VALUE_MG[6] = {82, 337, 365, 477, 1025, 0};
+const int16_t VALUE_EG[6] = {94, 281, 297, 512, 936, 0};
 
 // clang-format off
-const int PST_PAWN_MG[64] = {
+const int16_t PST_PAWN_MG[64] = {
       0,   0,   0,   0,   0,   0,  0,   0,
      98, 134,  61,  95,  68, 126, 34, -11,
      -6,   7,  26,  31,  65,  56, 25, -20,
@@ -19,7 +19,7 @@ const int PST_PAWN_MG[64] = {
       0,   0,   0,   0,   0,   0,  0,   0,
 };
 
-const int PST_PAWN_EG[64] = {
+const int16_t PST_PAWN_EG[64] = {
       0,   0,   0,   0,   0,   0,   0,   0,
     178, 173, 158, 134, 147, 132, 165, 187,
      94, 100,  85,  67,  56,  53,  82,  84,
@@ -30,7 +30,7 @@ const int PST_PAWN_EG[64] = {
       0,   0,   0,   0,   0,   0,   0,   0,
 };
 
-const int PST_KNIGHT_MG[64] = {
+const int16_t PST_KNIGHT_MG[64] = {
     -167, -89, -34, -49,  61, -97, -15, -107,
      -73, -41,  72,  36,  23,  62,   7,  -17,
      -47,  60,  37,  65,  84, 129,  73,   44,
@@ -41,7 +41,7 @@ const int PST_KNIGHT_MG[64] = {
     -105, -21, -58, -33, -17, -28, -19,  -23,
 };
 
-const int PST_KNIGHT_EG[64] = {
+const int16_t PST_KNIGHT_EG[64] = {
     -58, -38, -13, -28, -31, -27, -63, -99,
     -25,  -8, -25,  -2,  -9, -25, -24, -52,
     -24, -20,  10,   9,  -1,  -9, -19, -41,
@@ -52,7 +52,7 @@ const int PST_KNIGHT_EG[64] = {
     -29, -51, -23, -15, -22, -18, -50, -64,
 };
 
-const int PST_BISHOP_MG[64] = {
+const int16_t PST_BISHOP_MG[64] = {
     -29,   4, -82, -37, -25, -42,   7,  -8,
     -26,  16, -18, -13,  30,  59,  18, -47,
     -16,  37,  43,  40,  35,  50,  37,  -2,
@@ -63,7 +63,7 @@ const int PST_BISHOP_MG[64] = {
     -33,  -3, -14, -21, -13, -12, -39, -21,
 };
 
-const int PST_BISHOP_EG[64] = {
+const int16_t PST_BISHOP_EG[64] = {
     -14, -21, -11,  -8, -7,  -9, -17, -24,
      -8,  -4,   7, -12, -3, -13,  -4, -14,
       2,  -8,   0,  -1, -2,   6,   0,   4,
@@ -74,7 +74,7 @@ const int PST_BISHOP_EG[64] = {
     -23,  -9, -23,  -5, -9, -16,  -5, -17,
 };
 
-const int PST_ROOK_MG[64] = {
+const int16_t PST_ROOK_MG[64] = {
      32,  42,  32,  51, 63,  9,  31,  43,
      27,  32,  58,  62, 80, 67,  26,  44,
      -5,  19,  26,  36, 17, 45,  61,  16,
@@ -85,7 +85,7 @@ const int PST_ROOK_MG[64] = {
     -19, -13,   1,  17, 16,  7, -37, -26,
 };
 
-const int PST_ROOK_EG[64] = {
+const int16_t PST_ROOK_EG[64] = {
     13, 10, 18, 15, 12,  12,   8,   5,
     11, 13, 13, 11, -3,   3,   8,   3,
      7,  7,  7,  5,  4,  -3,  -5,  -3,
@@ -96,7 +96,7 @@ const int PST_ROOK_EG[64] = {
     -9,  2,  3, -1, -5, -13,   4, -20,
 };
 
-const int PST_QUEEN_MG[64] = {
+const int16_t PST_QUEEN_MG[64] = {
     -28,   0,  29,  12,  59,  44,  43,  45,
     -24, -39,  -5,   1, -16,  57,  28,  54,
     -13, -17,   7,   8,  29,  56,  47,  57,
@@ -107,7 +107,7 @@ const int PST_QUEEN_MG[64] = {
      -1, -18,  -9,  10, -15, -25, -31, -50,
 };
 
-const int PST_QUEEN_EG[64] = {
+const int16_t PST_QUEEN_EG[64] = {
      -9,  22,  22,  27,  27,  19,  10,  20,
     -17,  20,  32,  41,  58,  25,  30,   0,
     -20,   6,   9,  49,  47,  35,  19,   9,
@@ -118,7 +118,7 @@ const int PST_QUEEN_EG[64] = {
     -33, -28, -22, -43,  -5, -32, -20, -41,
 };
 
-const int PST_KING_MG[64] = {
+const int16_t PST_KING_MG[64] = {
     -65,  23,  16, -15, -56, -34,   2,  13,
      29,  -1, -20,  -7,  -8,  -4, -38, -29,
      -9,  24,   2, -16, -20,   6,  22, -22,
@@ -129,7 +129,7 @@ const int PST_KING_MG[64] = {
     -15,  36,  12, -54,   8, -28,  24,  14,
 };
 
-const int PST_KING_EG[64] = {
+const int16_t PST_KING_EG[64] = {
     -74, -35, -18, -18, -11,  15,   4, -17,
     -12,  17,  14,  17,  17,  38,  23,  11,
      10,  17,  23,  15,  20,  45,  44,  13,
@@ -141,16 +141,16 @@ const int PST_KING_EG[64] = {
 };
 // clang-format on
 
-const int *PST_MG[6] = {
+const int16_t *PST_MG[6] = {
     PST_PAWN_MG, PST_KNIGHT_MG, PST_BISHOP_MG, PST_ROOK_MG, PST_QUEEN_MG, PST_KING_MG,
 };
 
-const int *PST_EG[6] = {
+const int16_t *PST_EG[6] = {
     PST_PAWN_EG, PST_KNIGHT_EG, PST_BISHOP_EG, PST_ROOK_EG, PST_QUEEN_EG, PST_KING_EG,
 };
 
-int TABLE_MG[COLOR_COUNT][PIECE_COUNT][SQUARE_COUNT];
-int TABLE_EG[COLOR_COUNT][PIECE_COUNT][SQUARE_COUNT];
+int16_t TABLE_MG[COLOR_COUNT][PIECE_COUNT][SQUARE_COUNT];
+int16_t TABLE_EG[COLOR_COUNT][PIECE_COUNT][SQUARE_COUNT];
 
 void init_values(void) {
     for (Piece p = PAWN; p <= KING; p++) {
